update
html5页面优化

+++ bibliosoft/src/main/java/com/lib/bibliosoft/controller/ReaderController.java
@@ -358,11 +358,13 @@ public class ReaderController {
 
             //删除原图片
             Reader reader=readerRepository.findReaderByReaderId(readerId);
-            if(reader.getImgsrc()!=null){
+            if(reader.getImgsrc()!=null&&!reader.getImgsrc().equals("")){
                 String oldimgrc=reader.getImgsrc().substring(21);
                 //logger.info(upload.getAbsolutePath()+"\\"+oldimgrc);
-                File oldImg=new File(upload.getAbsolutePath()+"\\"+oldimgrc);
-                if(oldImg.delete()){logger.info("删除原图片成功");}else{logger.info("删除原图片失败");}
+                if(!oldimgrc.equals("defaultimg.jpg")){
+                    File oldImg=new File(upload.getAbsolutePath()+"\\"+oldimgrc);
+                    if(oldImg.delete()){logger.info("删除原图片成功");}else{logger.info("删除原图片失败");}
+                }
             }
             // 上传成功或者失败的提示
             String newfilename=FileNameUtil.getFileName(imgFile.getOriginalFilename());
@@ -418,16 +420,20 @@ public class ReaderController {
 
     @RequestMapping("/goSearch")
     public String goSearch(Model model,Integer booktypeid,Integer currpage) throws Exception{
-        if(booktypeid!=null){
+        if(booktypeid!=null&&currpage!=0){
             Integer totalPages =  bookSortService.PageBook(0,pagesize,booktypeid).getTotalPages();
             model.addAttribute("totalpages", totalPages);
             //获得每页的数据
             if(currpage == 1)
                 currpage = 0;
-            if(currpage == totalPages)
+            if(currpage == totalPages&&currpage>0)
                 currpage = totalPages-1;
             List<BookSort> bookSortList=bookSortService.PageBook(currpage,pagesize,booktypeid).getContent();
-            model.addAttribute("currpage",currpage+1);
+            if(totalPages>1){
+                model.addAttribute("currpage",currpage+1);
+            }else{
+                model.addAttribute("currpage",0);
+            }
             model.addAttribute("booklist",bookSortList);
             model.addAttribute("booktypeid",booktypeid);
             model.addAttribute("booktypelist",bookTypeRepository.findAll());
